Rename vector_int to use_vector
